include the schema rule to the migrations

+++ database/migrations/2021_02_08_232030_drop_bank_account_name_in_users_table.php
@@ -13,10 +13,12 @@ class DropBankAccountNameInUsersTable extends Migration
      */
     public function up()
     {
-        Schema::table('users', function (Blueprint $table) {
-            //
-            $table->dropColumn('bank_account_name');
-        });
+        if (!Schema::hasColumn('users', 'bank_account_name')){
+            Schema::table('users', function (Blueprint $table) {
+                //
+                $table->dropColumn('bank_account_name');
+            });
+        }
     }
 
     /**
